+++ moneta/www/static/css/forgot_pass.css
@@ -1,0 +1,140 @@
+body {
+    background-color: #F4DDF1;
+}
+.container-fluid{
+	font: 95% Arial, Helvetica, sans-serif;
+	max-width: 400px;
+	position: absolute;
+    top: 50%;
+    left: 50%;
+    margin: -150px 0 0 -150px;
+	padding: 30px;
+	background: #F7F7F7;
+}
+.container-fluid h1{
+	background: #D85258;
+	padding: 20px 8px;
+	font-size: 160%;
+	font-weight: 500;
+	text-align: center;
+	color: #fff;
+	margin: -16px -16px -16px -16px;
+}
+.container-fluid p{
+    text-align: center;
+	color: #555;
+    font:  20px Arial, Helvetica, sans-serif; ;
+
+}
+.container-fluid input[type="text"]{
+    display: inline-block;
+	-webkit-transition: all 0.30s ease-in-out;
+	-moz-transition: all 0.30s ease-in-out;
+	-ms-transition: all 0.30s ease-in-out;
+	-o-transition: all 0.30s ease-in-out;
+	outline: none;
+	box-sizing: border-box;
+	-webkit-box-sizing: border-box;
+	-moz-box-sizing: border-box;
+  	width: 100%;
+	background: #fff;
+	margin-bottom: 4%;
+	border: 1px solid #ccc;
+	padding: 3%;
+	color: #555;
+	font: 95% Arial, Helvetica, sans-serif;
+}
+.container-fluid input[type="text"]:focus{
+	box-shadow: 0 0 5px #D85258;
+	padding: 3%;
+	border: 1px solid #D85258;
+}
+
+.container-fluid input[type="submit"],
+.container-fluid input[type="button"]{
+	box-sizing: border-box;
+	-webkit-box-sizing: border-box;
+	-moz-box-sizing: border-box;
+	width: 100%;
+	padding: 3%;
+	background: #D85258;
+	border-bottom: 2px solid #D85258;
+	border-top-style: none;
+	border-right-style: none;
+	border-left-style: none;
+	color: #fff;
+}
+.container-fluid input[type="submit"]:hover,
+.container-fluid input[type="button"]:hover{
+	background: #D85258;}
+
+.container-for-not-user{
+    font: 95% Arial, Helvetica, sans-serif;
+	max-width: 400px;
+	position: absolute;
+    top: 50%;
+    left: 50%;
+    margin: -220px 0 0 -200px;
+	padding: 40px;
+	background: #F7F7F7;
+}
+.container-for-not-user h1{
+	background: #D85258;
+	padding: 20px 0;
+	font-size: 160%;
+	font-weight: 500;
+	text-align: center;
+	color: #fff;
+	margin: -16px -16px 16px -16px;
+}
+.container-for-not-user p{
+
+    text-align: center;
+	color: #555;
+    font:  20px Arial, Helvetica, sans-serif; ;
+
+}
+.container-for-not-user input[type="text"]
+{
+    display: inline-block;
+	-webkit-transition: all 0.30s ease-in-out;
+	-moz-transition: all 0.30s ease-in-out;
+	-ms-transition: all 0.30s ease-in-out;
+	-o-transition: all 0.30s ease-in-out;
+	outline: none;
+	box-sizing: border-box;
+	-webkit-box-sizing: border-box;
+	-moz-box-sizing: border-box;
+  	width: 100%;
+	background: #fff;
+	margin-bottom: 4%;
+	border: 1px solid #ccc;
+	padding: 3%;
+	color: #555;
+	font: 95% Arial, Helvetica, sans-serif;
+}
+.container-for-not-user input[type="text"]:focus
+{
+	box-shadow: 0 0 5px #D85258;
+
+	padding: 3%;
+	border: 1px solid #D85258;
+}
+
+.container-for-not-user input[type="submit"],
+.container-for-not-user input[type="button"]{
+	box-sizing: border-box;
+	-webkit-box-sizing: border-box;
+	-moz-box-sizing: border-box;
+	width: 100%;
+	padding: 3%;
+	background: #D85258;
+	border-bottom: 2px solid #D85258;
+	border-top-style: none;
+	border-right-style: none;
+	border-left-style: none;
+	color: #fff;
+}
+.container-for-not-user input[type="submit"]:hover,
+.container-for-not-user input[type="button"]:hover{
+	background: #D85258;}
