+++ moneta/www/templates/authentication/forgot_password.html
@@ -1,93 +1,17 @@
+{% load static %}
 <!DOCTYPE html>
 <html lang="en">
 <head>    <meta charset="UTF-8">    <title> Forgot password</title>
-
-    <style>
-body {
-    background-color: #E7C3AA;
-}
-.container-fluid{
-	font: 95% Arial, Helvetica, sans-serif;
-	max-width: 400px;
-	position: absolute;
-    top: 50%;
-    left: 50%;
-    margin: -150px 0 0 -150px;
-	padding: 30px;
-	background: #F7F7F7;
-}
-.container-fluid h1{
-	background: #D85258;
-	padding: 20px 0;
-	font-size: 160%;
-	font-weight: 500;
-	text-align: center;
-	color: #fff;
-	margin: -16px -16px 16px -16px;
-}
-.container-fluid p{
-    text-align: center;
-	color: #555;
-    font:  20px Arial, Helvetica, sans-serif; ;
-
-}
-.container-fluid input[type="text"]{
-    display: inline-block;
-	-webkit-transition: all 0.30s ease-in-out;
-	-moz-transition: all 0.30s ease-in-out;
-	-ms-transition: all 0.30s ease-in-out;
-	-o-transition: all 0.30s ease-in-out;
-	outline: none;
-	box-sizing: border-box;
-	-webkit-box-sizing: border-box;
-	-moz-box-sizing: border-box;
-  	width: 100%;
-	background: #fff;
-	margin-bottom: 4%;
-	border: 1px solid #ccc;
-	padding: 3%;
-	color: #555;
-	font: 95% Arial, Helvetica, sans-serif;
-}
-.container-fluid input[type="text"]:focus{
-	box-shadow: 0 0 5px #D85258;
-	padding: 3%;
-	border: 1px solid #D85258;
-}
-
-.container-fluid input[type="submit"],
-.container-fluid input[type="button"]{
-	box-sizing: border-box;
-	-webkit-box-sizing: border-box;
-	-moz-box-sizing: border-box;
-	width: 100%;
-	padding: 3%;
-	background: #D85258;
-	border-bottom: 2px solid #D85258;
-	border-top-style: none;
-	border-right-style: none;
-	border-left-style: none;
-	color: #fff;
-}
-.container-fluid input[type="submit"]:hover,
-.container-fluid input[type="button"]:hover{
-	background: #D85258;
-}
-    </style>
+<link rel="stylesheet" href="{% static 'css/forgot_pass.css' %}" type="text/css">
 </head>
 <body>
      <div class="container-fluid">
-<h1> Please, write your email! </h1>
-
+<h1>  Please, write your email!  </h1>
 <form method="POST"> {%csrf_token %}
-
-         <p> Email:  </p>   <input  type="text" name="email" >
-
-        <br>
-        <br>
-        <input  type="submit"  value="Submit">
-
-
+   <p> Email:  </p>   <input  type="text" name="email" >
+ <br>
+ <br>
+ <input  type="submit"  value="Submit">
 </form>
      </div>
 </body>
